% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dy.R
\name{dy}
\alias{dy}
\title{Funcion dy}
\usage{
dy(edo, x, y)
}
\arguments{
\item{edo}{La expresion que sera evaluada. DEBE escribirse con la funcion expression de R.}

\item{x}{El valor de x que sera computado en la funcion}

\item{y}{El valor de y que sera computado en la funcion}
}
\value{
Ninguno
}
\description{
Funcion que evalua una expresion dada, utilizando
los valores de "x" y "y" que recibe.
}
