% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEuler.R
\name{errorEuler}
\alias{errorEuler}
\title{Funcion errorEuler}
\usage{
errorEuler(en)
}
\arguments{
\item{en}{dataframe que contiene los resultados de la funcion metodoEuler}
}
\value{
Ninguno
}
\description{
Funcion que calcula los resultados exactos de una ecuacion diferencial
ordinaria de primer orden por medio del metodo de Euler disponible en
R, hayando el error absoluto a comparacion de los resultaods obtenidos
por medio de la funcion metodoEuler
}
