% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficarEuler.R
\name{graficarEuler}
\alias{graficarEuler}
\title{Funcion graficarEuler}
\usage{
graficarEuler(en, edo)
}
\arguments{
\item{en}{dataframe que contiene los resultados de la funcion metodoEuler}

\item{edo}{la expresion a evaluar por medio de la funcion dy}
}
\value{
Ninguno
}
\description{
Funcion que grafica la solucion de una ecuacion diferencial ordinaria
de primer orden, el campo vectorial en que esta inmersa, graficando los
puntos necesarios, y utiliza la funcion errorEuler para mostrar los
errores.Hace uso de una funcion embedida para graficar el campo vectorial.
}
