% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodoEuler.R
\name{metodoEuler}
\alias{metodoEuler}
\title{Funcion metodoEuler}
\usage{
metodoEuler(edo, h, x0, y0, xf)
}
\arguments{
\item{edo}{El nombre de la expresion a evaluar.}

\item{h}{El tamanio del paso entre los valores de x. NO MENORES A 1e-6}

\item{x0}{El valor inicial de x}

\item{y0}{El valor inicial de y al ser evaluada en x0}

\item{xf}{El valor final de x para el ejercicio}
}
\value{
dataframe con los valores calculados de x, acorde a h, y y
}
\description{
Funcion que resuelve ecuaciones diferenciales ordinarias
de primer orden, por medio del metodo numerico de Euler.
Esta funcion en particular UNICAMENTE puede utilizarse para
problemas de valor inicial.
}
